+++ app/src/main/java/com/tosslab/jandi/app/ui/maintab/file/FileListFragment.java
@@ -769,7 +769,7 @@ public class FileListFragment extends Fragment
         if (searchedFileItemListAdapter == null) {
             return;
         }
-        int fileId = event.getFileId();
+        long fileId = event.getFileId();
         int position = searchedFileItemListAdapter.findPositionByFileId(fileId);
         if (position < 0) {
             return;
